Fix algorithms folder name typo
