+++ algorithms/Searching/binary_search.py
