+++ algorithms/Searching/linear_search.py
