+++ algorithms/ciphers/xor_cipher.cpp
